Extract gameplay route helpers
